DDF-04485 Migrating to React
